+++ src/main/java/domain/engine/ImmovableEngine.java
@@ -1,4 +1,4 @@
-package utils;
+package domain.engine;
 
 public class ImmovableEngine implements Engine {
     @Override
